FEATURE: insirido metodo getmapping que simula dos os requisitos do teste

+++ src/main/java/maymb/teste_pratico_iniflex/dto/FuncionarioRequestDTO.java
@@ -1,11 +1,12 @@
 package maymb.teste_pratico_iniflex.dto;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record FuncionarioRequestDTO(
         String nome,
         LocalDate nascimento,
-        Double salario,
+        BigDecimal salario,
         String funcao
 
 ) {
